pixly/pixly-backend/requirements.txt
blinker==1.6.3
boto3==1.28.62
botocore==1.31.62
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
Pillow==10.0.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==3.0.0
